Almost a working state?
Error:
[ERROR] testSaveMethod  Time elapsed: 0.037 s  <<< ERROR!
java.lang.IllegalArgumentException: Class class org.mskcc.cmo.metadb.model.MetaDbSample is not a valid entity class. Please check the entity mapping.
	at org.mskcc.cmo.metadb.persistence.MockNeo4jDbTest.testSaveMethod(MockNeo4jDbTest.java:109)

Want to try loading an actual request JSON into the test database and seeing if the error we are seeing might
be due to the relationships not all being mapped out when we attempt to just persist a sample

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>